修改购买数量的类型

+++ onl-api/src/main/java/rebue/onl/to/ModifySaleCountByIdTo.java
@@ -1,11 +1,13 @@
 package rebue.onl.to;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class ModifySaleCountByIdTo {
 
-	private Long id;
-	
-	private Integer buyCount;
+    private Long id;
+
+    private BigDecimal buyCount;
 }
